personal_finance_agent/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0

# Data Processing & Analytics
pandas>=2.1.0
plotly>=5.17.0
matplotlib>=3.7.0

# Web Interface
streamlit>=1.28.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Authentication & Security
authlib>=1.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0

# Scheduling & State Management
apscheduler>=3.10.4
transitions>=0.9.0

# Additional utilities
requests>=2.31.0
python-multipart>=0.0.6
aiofiles>=23.2.1 